# [3.10] Cache construction of middleware handlers