--- a/aiohttp/web_app.py
+++ b/aiohttp/web_app.py
@@ -89,6 +89,7 @@ class Application(MutableMapping[Union[str, AppKey[Any]], Any]):
             "_handler_args",
             "_middlewares",
             "_middlewares_handlers",
+            "_middleware_cache",
             "_run_middlewares",
             "_state",
             "_frozen",
@@ -143,6 +144,10 @@ class Application(MutableMapping[Union[str, AppKey[Any]], Any]):
         self._middlewares_handlers: _MiddlewaresHandlers = None
         # initialized on freezing
         self._run_middlewares: Optional[bool] = None
+        #
+        self._middleware_cache: Dict[
+            Tuple[int, Callable[[Request], Awaitable[StreamResponse]], Tuple[int, ...]]
+        ] = {}
 
         self._state: Dict[Union[AppKey[Any], str], object] = {}
         self._frozen = False
@@ -525,18 +530,40 @@ class Application(MutableMapping[Union[str, AppKey[Any]], Any]):
             handler = match_info.handler
 
             if self._run_middlewares:
-                for app in match_info.apps[::-1]:
-                    for m, new_style in app._middlewares_handlers:  # type: ignore[union-attr]
-                        if new_style:
-                            handler = update_wrapper(
-                                partial(m, handler=handler), handler  # type: ignore[misc]
-                            )
-                        else:
-                            handler = await m(app, handler)  # type: ignore[arg-type,assignment]
+                handler = await self._apply_middlewares(
+                    self, handler, match_info.apps[::-1]
+                )
 
             resp = await handler(request)
 
         return resp
+
+    async def _apply_middlewares(
+        self,
+        app: "Application",
+        handler: Callable[[Request], Awaitable[StreamResponse]],
+        apps: Tuple["Application", ...],
+    ) -> Callable[[Request], Awaitable[StreamResponse]]:
+        """Apply middlewares to handler."""
+        cache_key = (id(app), handler, tuple(id(app) for app in apps))
+        if cache_key in self._middleware_cache:
+            return self._middleware_cache[cache_key]
+
+        can_cache: bool = True
+        for app in apps:
+            for m, new_style in app._middlewares_handlers:  # type: ignore[union-attr]
+                if new_style:
+                    handler = update_wrapper(
+                        partial(m, handler=handler), handler  # type: ignore[misc]
+                    )
+                else:
+                    handler = await m(app, handler)  # type: ignore[arg-type,assignment]
+                    can_cache = False
+
+        if can_cache:
+            self._middleware_cache[cache_key] = handler
+
+        return handler
 
     def __call__(self) -> "Application":
         """gunicorn compatibility"""