--- a/aiohttp/web_app.py
+++ b/aiohttp/web_app.py
@@ -1,7 +1,7 @@
 import asyncio
 import logging
 import warnings
-from functools import partial, update_wrapper
+from functools import cache, partial, update_wrapper
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -230,6 +230,9 @@ class Application(MutableMapping[Union[str, AppKey[Any]], Any]):
     def __iter__(self) -> Iterator[Union[str, AppKey[Any]]]:
         return iter(self._state)
 
+    def __hash__(self) -> int:
+        return id(self)
+
     @overload  # type: ignore[override]
     def get(self, key: AppKey[_T], default: None = ...) -> Optional[_T]: ...
 
@@ -286,6 +289,9 @@ class Application(MutableMapping[Union[str, AppKey[Any]], Any]):
         self._on_shutdown.freeze()
         self._on_cleanup.freeze()
         self._middlewares_handlers = tuple(self._prepare_middleware())
+        self._has_legacy_middlewares = any(
+            not new_style for _, new_style in self._middlewares_handlers
+        )
 
         # If current app and any subapp do not have middlewares avoid run all
         # of the code footprint that it implies, which have a middleware
@@ -527,37 +533,45 @@ class Application(MutableMapping[Union[str, AppKey[Any]], Any]):
             handler = match_info.handler
 
             if self._run_middlewares:
-                handler = await self._apply_middlewares(handler, match_info.apps[::-1])
+                if self._has_legacy_middlewares:
+                    # Slow path for legacy handlers
+                    handler = await self._apply_middlewares(
+                        handler, match_info.apps[::-1]
+                    )
+                else:
+                    handler = Application._build_middlewares(
+                        handler, match_info.apps[::-1]
+                    )
 
             resp = await handler(request)
 
         return resp
 
     async def _apply_middlewares(
-        self,
-        handler: Handler,
-        apps: Tuple["Application", ...],
+        self, handler: Handler, apps: Tuple["Application", ...]
     ) -> Callable[[Request], Awaitable[StreamResponse]]:
-        """Apply middlewares to handler."""
-        cache_key = (handler, tuple(id(app) for app in apps))
-
-        if cache_key in self._middlewares_cache:
-            return self._middlewares_cache[cache_key]
-
-        can_cache: bool = True
+        """Apply middlewares to handler with support for legacy handlers."""
         for app in apps:
-            for m, new_style in app._middlewares_handlers:  # type: ignore[union-attr]
+            for m, new_style in app._middlewares_handlers:
                 if new_style:
                     handler = update_wrapper(
                         partial(m, handler=handler), handler  # type: ignore[misc]
                     )
                 else:
-                    handler = await m(app, handler)  # type: ignore[arg-type,assignment]
-                    can_cache = False
+                    handler = await m(app, handler)
 
-        if can_cache:
-            self._middlewares_cache[cache_key] = handler
+        return handler
 
+    @staticmethod
+    @cache
+    def _build_middlewares(
+        handler: Handler,
+        apps: Tuple["Application", ...],
+    ) -> Callable[[Request], Awaitable[StreamResponse]]:
+        """Apply middlewares to handler."""
+        for app in apps:
+            for m, _ in app._middlewares_handlers:  # type: ignore[union-attr]
+                handler = update_wrapper(partial(m, handler=handler), handler)  # type: ignore[misc]
         return handler
 
     def __call__(self) -> "Application":