Will do a PR to master once I'm happy with this, but since I'm traveling, I want to test live for now.  master will be much simpler since we don't have to deal with legacy middleware.
master version is https://github.com/aio-libs/aiohttp/pull/9158


<!-- Thank you for your contribution! -->

## What do these changes do?

Cache construction of middleware.  Calling `update_wrapper` many times for each request with middleware is a more expensive than running all the middlewares 

## Are there changes in behavior for the user?

No

## Is it a substantial burden for the maintainers to support this?
no

Before
```
bdraco@MacBook-Pro-5 ~ % wrk -t 10 http://localhost:8123
Running 10s test @ http://localhost:8123
  10 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     6.83ms   33.14ms 348.84ms   96.25%
    Req/Sec     1.89k   799.64     2.41k    80.83%
  181206 requests in 10.01s, 0.95GB read
Requests/sec:  18100.19
Transfer/sec:     97.39MB
```



After
```
bdraco@MacBook-Pro-5 ~ % wrk -t 10 http://localhost:8123
Running 10s test @ http://localhost:8123
  10 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   386.34us  124.22us   5.11ms   98.73%
    Req/Sec     2.63k   156.95     2.77k    96.04%
  264316 requests in 10.10s, 1.39GB read
Requests/sec:  26169.50
Transfer/sec:    140.81MB
```

![middleware](https://github.com/user-attachments/assets/dd49750d-48b4-49eb-bfe9-39a3561934ff)