@bdraco 2024-09-16
> Why would this be 3.10 only?

master version is https://github.com/aio-libs/aiohttp/pull/9158

I'm testing on 3.10 and since I'm traveling it was easier to do 3.10 first and work backwards